ypm/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.7.1
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.4.0
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jira==3.8.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.7
langchain-core==0.3.16
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.2
langsmith==0.1.142
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.11
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
